Move glob utils to different utils file
*  Move glob utils function into a path utils file
*  Remove commented out code
